+++ upkie/utils/exceptions.py
@@ -15,3 +15,6 @@ class FallDetected(UpkieException):
 
 class ModelError(UpkieException):
     """Error related to the robot model."""
+
+class ActionError(UpkieException):
+    """Error related to an action supplied to an Upkie environment."""