+++ upkie/__init__.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2023 Inria
 # SPDX-License-Identifier: Apache-2.0
+# Copyright 2023 Inria
 
 """Python module to control Upkie wheeled bipeds."""
 
-__version__ = "3.0.0"
+__version__ = "3.1.0"