+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename ``start_pid_balancer.sh`` to ``try_pid_balancer.sh``
+
 ## [3.0.0] - 2023-12-01
 
 ### Added