+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- ModelError exception
+
 ### Changed
 
 - **Breaking:** Rename info key to `"spine_observation"` in all environments