+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking:** Rename info key to `"spine_observation"` in all environments
 - **Breaking:** Switch `UpkieServos` env to dictionary action and observation
 - Rename ``start_pid_balancer.sh`` to ``try_pid_balancer.sh``
+- envs: Make parsing of first observation optional for non-base environments
 
 ## [3.0.0] - 2023-12-01