+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ModelError exception
+- utils: ActionError exception
+- utils: ModelError exception
 
 ### Changed