+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2023-12-22
+
 ### Added
 
 - examples: Wheeled balancing with wheel torque control
@@ -15,10 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **Breaking:** Rename info key to `"spine_observation"` in all environments
-- **Breaking:** Switch `UpkieServos` env to dictionary action and observation
 - Rename ``start_pid_balancer.sh`` to ``try_pid_balancer.sh``
 - envs: Make parsing of first observation optional for non-base environments
+- envs: Rename info key to `"spine_observation"` in all environments
+- envs: Switch `UpkieServos` env to dictionary action and observation
 
 ## [3.0.0] - 2023-12-01
 
@@ -380,7 +382,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Starting this changelog.
 
-[unreleased]: https://github.com/upkie/upkie/compare/v3.0.0...HEAD
+[unreleased]: https://github.com/upkie/upkie/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/upkie/upkie/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/upkie/upkie/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/upkie/upkie/compare/v1.5.0...v2.0.0
 [1.5.0]: https://github.com/upkie/upkie/compare/v1.4.0...v1.5.0